requirements.txt
mcstatus
discord.py
ping3
requests
load_dotenv
pytz